[refactor](session) Keep both ProtocolAdapter implementations in a protocol subpackage
qe.protocol defines the interface; each front end implements it in its own
protocol subpackage: mysql.protocol.MysqlProtocolAdapter and now
service.arrowflight.protocol.FlightProtocolAdapter. The result senders of the
next step land in the same two places.
